chore: move class to file
